requirements.txt

# Core scientific packages
numpy>=1.20.0
matplotlib>=3.5.0,<3.10.0
pandas>=1.3.0

# Deep Learning
torch>=2.0.0
torchvision>=0.15.0

# Reinforcement learning environment
gymnasium>=0.28.1

# Interactive widgets and progress bars (for notebooks)
ipywidgets>=7.6.0
tqdm>=4.62.0

# IPython display components (for notebook visualizations)
ipython>=7.0.0

# Configuration management
pyyaml>=6.0

# Logging and experiment tracking
tensorboard>=2.11.0

# Type hints
typing-extensions>=4.0.0

# Testing (optional)
pytest>=7.0.0
pytest-cov>=3.0.0